+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## unreleased
 * Add `acquirerReferenceNumber` to `Transaction`
+* Add `billingAgreementId` to `PayPalDetails`
 
 ## 3.2.0
 * Deprecate `MasterpassCard` payment method